[mlir] correct handling of fresh alias sets
Previously, dataflow initialization and transfer functions would
allocate a fresh alias class for allocation-like cases that are known
not to alias with anything, and do so _on every invocation_. This is
unsound in the general case as the class may keep changing, which could
affect convergence to fix point. Instead, add a cache that associates a
class with a value on the first invocation of the dataflow-related
functions and use the cached value on each subsequent invocation.

Fix a couple of issues unconvered due to the change:
  - not recreate the common scope that function results may be pointing
    to for each capturable operand;
  - invert the condition of handling nocapture in function results;
  - add a TODO for better handling of alias classes of function results
    in presence of function-level attributes.