Reimplement MLIR tangent for scf.for generically (#1649)
* Reimplement MLIR tangent for scf.for generically

Introduce an interface for ops with region control flow and use that
interface inside the tangent implementation for scf.for. This interface
can also work for other kinds of control flow ops as long as they don't
transfer control flow between regions (e.g. like scf.while does).

Further cleanup is necessary to clearly delegate implementation of
AutoDiffOpInterface to this new interface for any compatible op via
traits.

* Generalize

* fixup

* Decrease the amount of tablegen'd code

* Add some documentation and use better types

* add forward diff support for affine.for and affine.if

* bazel

* add forward diff rule for scf.while

Fix a minor bug discovered in the process: the interface was designed
for constant folding and expects a list of null attributes of the
appropriate size...

* add executeop

* add cf

* fixup

---------

Co-authored-by: William S. Moses <gh@wsmoses.com>